add Intafeces

+++ src/main/java/com/antoniofrische/bestgamevendor/repositorios/IUserRepository.java
@@ -4,7 +4,7 @@ import com.antoniofrische.bestgamevendor.models.UsuarioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface IUserRepository extends JpaRepository<UsuarioEntity, Integer> {
+public interface IUserRepository extends JpaRepository<UsuarioEntity, Long> {
     @Query("SELECT u FROM UsuarioEntity u WHERE u.email = ?1")
     public UsuarioEntity findByEmail(String email);
